Moved data to sep file + start replace routes
